requirement.txt
aniso8601==0.82
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
cssselect==1.0.3
Flask==0.10.1
Flask-RESTful==0.2.12
goose-extractor==1.0.25
gunicorn==19.0.0
idna==2.6
itsdangerous==0.24
jieba==0.39
Jinja2==2.7.3
JPype1==0.6.3
justext
lxml
MarkupSafe==0.23
nltk==3.3
Pillow==5.1.0
PySocks==1.6.8
pytz==2014.4
readability-lxml==0.7
requests==2.18.4
six==1.7.2
urllib3==1.22
Werkzeug==0.9.6
